LPS-124961 Class that implements 'ConfigurationBeanDeclaration' should be in package *.settings.definition
