--- modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/configuration/definition/CPOptionConfigurationBeanDeclaration.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.commerce.product.internal.configuration.definition;
-
-import com.liferay.commerce.product.configuration.CPOptionConfiguration;
-import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Alessio Antonio Rendina
- */
-@Component(enabled = false, service = ConfigurationBeanDeclaration.class)
-public class CPOptionConfigurationBeanDeclaration
-	implements ConfigurationBeanDeclaration {
-
-	@Override
-	public Class<?> getConfigurationBeanClass() {
-		return CPOptionConfiguration.class;
-	}
-
-}
